feat: update agent metrics to include RAM available and use float for CPU usage

+++ common/pkg/model/agent.go
@@ -3,7 +3,7 @@ package model
 type Agent struct {
 	AgentID      string  `json:"agent_id"`
 	Online       bool    `json:"online"`
-	CPUUsage     uint32  `json:"cpu_usage"`
+	CPUUsage     float32 `json:"cpu_usage"`
 	RAMTotal     uint64  `json:"ram_total"`
 	RAMAvailable uint64  `json:"ram_available"`
 	Uptime       float64 `json:"uptime"`
